// qp-explorer-frontend/src/typechain-types/OperatorRelation.d.ts
interface OperatorRelationInterface extends ethers.utils.Interface {
  functions: {
    "assignOperator(address)": FunctionFragment;
    "delegateLookup(address)": FunctionFragment;
    "getDelegateForOperator(address)": FunctionFragment;
    "nodeOperator(address)": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "assignOperator",
    values: [string]
  ): string;
  encodeFunctionData(
    functionFragment: "delegateLookup",
    values: [string]
  ): string;
  encodeFunctionData(
    functionFragment: "getDelegateForOperator",
    values: [string]
  ): string;
  encodeFunctionData(
    functionFragment: "nodeOperator",
    values: [string]
  ): string;

  decodeFunctionResult(
    functionFragment: "assignOperator",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "delegateLookup",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getDelegateForOperator",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "nodeOperator",
    data: BytesLike
  ): Result;

  events: {
    "NodeOperatorAssigned(address,address)": EventFragment;
  };

  getEvent(nameOrSignatureOrTopic: "NodeOperatorAssigned"): EventFragment;
}

export type NodeOperatorAssignedEvent = TypedEvent<
  [string, string] & { delegate: string; nodeOperator: string }
>;

export class OperatorRelation extends BaseContract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  listeners<EventArgsArray extends Array<any>, EventArgsObject>(
    eventFilter?: TypedEventFilter<EventArgsArray, EventArgsObject>
  ): Array<TypedListener<EventArgsArray, EventArgsObject>>;
  off<EventArgsArray extends Array<any>, EventArgsObject>(
    eventFilter: TypedEventFilter<EventArgsArray, EventArgsObject>,
    listener: TypedListener<EventArgsArray, EventArgsObject>
  ): this;
  on<EventArgsArray extends Array<any>, EventArgsObject>(
    eventFilter: TypedEventFilter<EventArgsArray, EventArgsObject>,
    listener: TypedListener<EventArgsArray, EventArgsObject>
  ): this;
  once<EventArgsArray extends Array<any>, EventArgsObject>(
    eventFilter: TypedEventFilter<EventArgsArray, EventArgsObject>,
    listener: TypedListener<EventArgsArray, EventArgsObject>
  ): this;
  removeListener<EventArgsArray extends Array<any>, EventArgsObject>(
    eventFilter: TypedEventFilter<EventArgsArray, EventArgsObject>,
    listener: TypedListener<EventArgsArray, EventArgsObject>
  ): this;
  removeAllListeners<EventArgsArray extends Array<any>, EventArgsObject>(
    eventFilter: TypedEventFilter<EventArgsArray, EventArgsObject>
  ): this;

  listeners(eventName?: string): Array<Listener>;
  off(eventName: string, listener: Listener): this;
  on(eventName: string, listener: Listener): this;
  once(eventName: string, listener: Listener): this;
  removeListener(eventName: string, listener: Listener): this;
  removeAllListeners(eventName?: string): this;

  queryFilter<EventArgsArray extends Array<any>, EventArgsObject>(
    event: TypedEventFilter<EventArgsArray, EventArgsObject>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEvent<EventArgsArray & EventArgsObject>>>;

  interface: OperatorRelationInterface;

  functions: {
    assignOperator(
      toOp: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<ContractTransaction>;

    delegateLookup(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<[string, number] & { delegate: string; deleted: number }>;

    getDelegateForOperator(
      operator: string,
      overrides?: CallOverrides
    ): Promise<[[string, number] & { delegate: string; deleted: number }]>;

    nodeOperator(arg0: string, overrides?: CallOverrides): Promise<[string]>;
  };

  assignOperator(
    toOp: string,
    overrides?: Overrides & { from?: string | Promise<string> }
  ): Promise<ContractTransaction>;

  delegateLookup(
    arg0: string,
    overrides?: CallOverrides
  ): Promise<[string, number] & { delegate: string; deleted: number }>;

  getDelegateForOperator(
    operator: string,
    overrides?: CallOverrides
  ): Promise<[string, number] & { delegate: string; deleted: number }>;

  nodeOperator(arg0: string, overrides?: CallOverrides): Promise<string>;

  callStatic: {
    assignOperator(toOp: string, overrides?: CallOverrides): Promise<void>;

    delegateLookup(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<[string, number] & { delegate: string; deleted: number }>;

    getDelegateForOperator(
      operator: string,
      overrides?: CallOverrides
    ): Promise<[string, number] & { delegate: string; deleted: number }>;

    nodeOperator(arg0: string, overrides?: CallOverrides): Promise<string>;
  };

  filters: {
    "NodeOperatorAssigned(address,address)"(
      delegate?: null,
      nodeOperator?: null
    ): TypedEventFilter<
      [string, string],
      { delegate: string; nodeOperator: string }
    >;

    NodeOperatorAssigned(
      delegate?: null,
      nodeOperator?: null
    ): TypedEventFilter<
      [string, string],
      { delegate: string; nodeOperator: string }
    >;
  };

  estimateGas: {
    assignOperator(
      toOp: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<BigNumber>;

    delegateLookup(arg0: string, overrides?: CallOverrides): Promise<BigNumber>;

    getDelegateForOperator(
      operator: string,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    nodeOperator(arg0: string, overrides?: CallOverrides): Promise<BigNumber>;
  };

  populateTransaction: {
    assignOperator(
      toOp: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<PopulatedTransaction>;

    delegateLookup(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    getDelegateForOperator(
      operator: string,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    nodeOperator(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;
  };
}